some clean up for reddit widget
