+++ src/components/Reddit/index.jsx
@@ -1,114 +1,83 @@
 import React, { Component } from 'react';
-import Divider from 'material-ui/Divider';
 import { List, ListItem } from 'material-ui/List';
 import NavigationRefresh from 'material-ui/svg-icons/navigation/refresh';
-import IconButton from 'material-ui/IconButton';
-import FlatButton from 'material-ui/FlatButton';
+import { IconButton, FlatButton, Divider, SelectField, MenuItem } from 'material-ui';
 import LocalStorageAPI from './../../helpers/localstorageAPI';
+import styles from './styles';
 /* global chrome */
 /* global navigator */
 
 const Snoowrap = require('snoowrap');
 
-const styles = {
-  primaryText: {
-    display: 'block',
-    marginLeft: '50px',
-    color: '#00bcd4',
-  },
-  loginButton: {
-    marginTop: '5px',
-  },
-  headerBar: {
-    height: '40px',
-  },
-  refreshIconButton: {
-    position: 'absolute',
-    top: '0px',
-    right: '25px',
-    color: 'gray',
-  },
-  refreshIcon: {
-    color: 'gray',
-    height: '30px',
-    width: '30px',
-  },
-  secondaryText: {
-    display: 'block',
-    marginLeft: '50px',
-  },
-  img: {
-    maxWidth: '100%',
-  },
-  leftIcon: {
-    height: 'auto',
-    maxHeight: '80px',
-    maxWidth: '100px',
-    width: 'auto',
-    margin: 'auto',
-    padding: 'auto',
-  },
-  tabHeadline: {
-    fontSize: 12,
-    paddingTop: 2,
-    marginBottom: 2,
-    fontWeight: 200,
-  },
-  listStyle: {
-    listStyleType: 'none',
-    margin: 0,
-    padding: 0,
-  },
-  divStyle: {
-    display: 'flex',
-    flexDirection: 'column',
-    background: '#75a3a3',
-    width: '100%',
-    height: '100%',
-  },
-  expand: {
-    width: '100%',
-    height: 'calc(100% - 40px)',
-    maxHeight: 'calc(100% - 40px)',
-    overflow: 'auto',
-  },
-  expand2: {
-    width: '100%',
-    height: '100%',
-    maxHeight: '100%',
-  },
-  listItemStyle: {
-    minHeight: '100px',
-    maxHeight: 'auto',
-  },
-};
+const countItems = [];
+for (let i = 50; i <= 400; i += 50) {
+  countItems.push(<MenuItem value={i} key={i} primaryText={`${i} Posts`} />);
+}
+
+function mediaToDisplay(post) {
+  const expression = '^https?://.*(.jpg)$';
+  const regex = new RegExp(expression);
+  if (post.media != null) {
+    if (post.is_reddit_media_domain) {
+      return (
+        <iframe
+          src={post.media.reddit_video.fallback_url}
+          title={post.title}
+          style={styles.img}
+        />
+      );
+    }
+    return (
+      <div
+        dangerouslySetInnerHTML={{ __html: post.media_embed.content }}
+        style={styles.img}
+      />
+    );
+  }
+  if (post.url.match(regex)) {
+    return <img src={post.url} alt={post.url} style={styles.img} />;
+  }
+  return null;
+}
 
 class Reddit extends Component {
   constructor(props) {
     super(props);
     this.baseUrl = 'https://www.reddit.com';
-    this.state = { accessToken: null, hotList: [], storageKey: this.constructor.name };
-    const tokenObject = LocalStorageAPI.get(this.state.storageKey);
-    if (tokenObject != null) {
-      this.state.accessToken = tokenObject.AuthToken;
-      this.getRedditHot();
-    }
+    this.state = {
+      accessToken: null,
+      hotList: [],
+      storageKey: this.constructor.name,
+      countValue: 50,
+    };
+
     this.login = this.login.bind(this);
     this.getRedditHot = this.getRedditHot.bind(this);
+    this.handleChange = this.handleChange.bind(this);
+
+    const localStorageState = LocalStorageAPI.get(this.state.storageKey);
+    if (localStorageState != null) {
+      this.state = localStorageState;
+      this.getRedditHot();
+    }
+  }
+  componentDidUpdate() {
+    LocalStorageAPI.put(this.state.storageKey, this.state);
   }
 
-  getRedditHot() {
+  getRedditHot(limitCount) {
+    const hotLimit = limitCount || this.state.countValue;
     const token = this.state.accessToken;
+    if (!token) {
+      return;
+    }
     const r = new Snoowrap({
       userAgent: navigator.userAgent,
       accessToken: token,
     });
-    console.log('getting list');
-    r.getHot({ limit: 100 }).then(posts => this.setState({ hotList: posts }));
-    console.log(this.state.hotList);
+    r.getHot({ limit: hotLimit }).then(posts => this.setState({ hotList: posts }));
   }
   redditList(post) {
-    console.log(post);
     const expression = '^https?://.*(.jpg)$';
     const regex = new RegExp(expression);
 
@@ -127,7 +96,7 @@ class Reddit extends Component {
             post.media != null ||
             post.is_reddit_media_domain ? null : <div />
           }
-          nestedItems={[this.mediaToDisplay(post)]}
+          nestedItems={[mediaToDisplay(post)]}
         />
         <Divider />
       </div>
@@ -141,46 +110,27 @@ class Reddit extends Component {
       { url: authUrl, interactive: true },
       (responseUrl) => {
         const token = responseUrl.match(/#(?:access_token)=([\S\s]*?)&/)[1];
-        console.log(token);
-        LocalStorageAPI.put(this.state.storageKey, { AuthToken: token });
         this.setState({ accessToken: token });
+        this.getRedditHot();
       },
     );
   }
-
-  mediaToDisplay(post) {
-    const token = this.state.accessToken;
-    const expression = '^https?://.*(.jpg)$';
-    const regex = new RegExp(expression);
-    if (post.media != null) {
-      if (post.is_reddit_media_domain) {
-        console.log('returned reddit media');
-        return (
-          <iframe
-            src={post.media.reddit_video.fallback_url}
-            title={post.title}
-            style={styles.img}
-          />
-        );
-      }
-      return (
-        <div
-          dangerouslySetInnerHTML={{ __html: post.media_embed.content }}
-          style={styles.img}
-        />
-      );
-    }
-    if (post.url.match(regex)) {
-      return <img src={post.url} alt={post.url} style={styles.img} />;
-    }
-    return null;
+  handleChange(event, index, value) {
+    this.getRedditHot(value);
+    this.setState({ countValue: value });
   }
-
   render() {
     return (
       <div style={styles.expand2}>
         <div style={styles.headerBar}>
           <FlatButton onClick={this.login} label="login" style={styles.loginButton} />
+          <SelectField
+            value={this.state.countValue}
+            onChange={this.handleChange}
+            style={styles.selectField}
+          >
+            {countItems}
+          </SelectField>
           <IconButton style={styles.refreshIconButton}>
             <NavigationRefresh onClick={this.getRedditHot} color="gray" />
           </IconButton>
