+++ src/components/Reddit/styles.jsx
@@ -1,0 +1,80 @@
+const styles = {
+  primaryText: {
+    display: 'block',
+    marginLeft: '50px',
+    color: '#00bcd4',
+  },
+  loginButton: {
+    marginTop: '5px',
+  },
+  headerBar: {
+    height: '40px',
+  },
+  refreshIconButton: {
+    position: 'absolute',
+    top: '0px',
+    right: '25px',
+    color: 'gray',
+  },
+  refreshIcon: {
+    color: 'gray',
+    height: '30px',
+    width: '30px',
+  },
+  secondaryText: {
+    display: 'block',
+    marginLeft: '50px',
+  },
+  img: {
+    maxWidth: '100%',
+  },
+  leftIcon: {
+    height: 'auto',
+    maxHeight: '80px',
+    maxWidth: '100px',
+    width: 'auto',
+    margin: 'auto',
+    padding: 'auto',
+  },
+  tabHeadline: {
+    fontSize: 12,
+    paddingTop: 2,
+    marginBottom: 2,
+    fontWeight: 200,
+  },
+  listStyle: {
+    listStyleType: 'none',
+    margin: 0,
+    padding: 0,
+  },
+  divStyle: {
+    display: 'flex',
+    flexDirection: 'column',
+    background: '#75a3a3',
+    width: '100%',
+    height: '100%',
+  },
+  expand: {
+    width: '100%',
+    height: 'calc(100% - 40px)',
+    maxHeight: 'calc(100% - 40px)',
+    overflow: 'auto',
+  },
+  expand2: {
+    width: '100%',
+    height: '100%',
+    maxHeight: '100%',
+  },
+  listItemStyle: {
+    minHeight: '100px',
+    maxHeight: 'auto',
+  },
+  selectField: {
+    width: '150px',
+    position: 'absolute',
+    top: '0px',
+    right: '75px',
+  },
+};
+
+export default styles;
